MedAI/requirements.txt
altair==4.1.0
anyio==3.4.0
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
catalogue==2.0.6
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
decorator==5.1.0
dill==0.3.4
entrypoints==0.3
fastapi==0.70.1
fastcore==1.3.27
flatbuffers==2.0
fonttools==4.28.3
gunicorn==20.1.0
h11==0.12.0
idna==3.3
imageio==2.13.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipython==7.30.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.2.1
kiwisolver==1.3.2
loguru==0.5.3
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
networkx==2.6.3
numpy
onnx
onnxruntime==1.12.0
opencv-python==4.5.4.60
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.24
protobuf==3.19.4
ptyprocess==0.7.0
pydantic==1.8.2
pydicom==2.2.2
Pygments==2.10.0
pylibjpeg==1.3.0
pylibjpeg-libjpeg==1.2.0
pylibjpeg-openjpeg==1.2.0
pylibjpeg-rle==1.2.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
rich==10.16.1
scikit-image==0.19.0
scipy==1.7.3
selectivesearch==0.4
six==1.16.0
sniffio==1.2.0
srsly==2.4.2
starlette==0.16.0
tifffile==2021.11.2
toolz==0.11.2
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
uvicorn==0.18.0
wasabi==0.9.0
wcwidth==0.2.5
zipp==3.6.0
pyproject
pycocotools
